requirements.txt
python-telegram-bot
requests
argparse